fix(next): updating next version (#25)
Next builds are getting locked up on the CI server, and some
say that the latest version of NextJS solves this issue. In addition,
this also introduces a Yarn offline mirror in hopes to reduce the
overall build time.

After some investigation, it appears that 9.3.4 and above have
memory issues with `next build`. Seeing if downgrading will solve
the exit code 137 issues on Jenkins.

Also updated:
- Included `fast-sass-loader` for dev environments
- Swapped out `next-progressbar` with `webpackbar`
- added linting rule for `components` folder