chore(dependabot): upgrading commitlint and sub-dependency `dot-prop`
Dependabot notified of a security risk with versions of `dot-prop`
before v5.1.1. This is a dependency of commitlint, which the template
was running an older version of. While the latest version of commitlint
requires `dot-prop` version `5.1.0`, for now we will need to set a
resolution in our `package.json` until commitlint will upgrade the
dependency in their project.